# Use gzip level 6 for faster json compression