--- a/multiqc/report.py
+++ b/multiqc/report.py
@@ -795,7 +795,7 @@ def compress_json(data):
     # Stream to an in-memory buffer rather than compressing the big string
     # at once. This saves memory.
     buffer = io.BytesIO()
-    with gzip.open(buffer, "wt", encoding="utf-8", compresslevel=9) as gzip_buffer:
+    with gzip.open(buffer, "wt", encoding="utf-8", compresslevel=6) as gzip_buffer:
         dump_json(data, gzip_buffer)
     base64_bytes = base64.b64encode(buffer.getvalue())
     return base64_bytes.decode("ascii")