--- a/multiqc/report.py
+++ b/multiqc/report.py
@@ -796,6 +796,7 @@ def compress_json(data):
     # at once. This saves memory.
     buffer = io.BytesIO()
     with gzip.open(buffer, "wt", encoding="utf-8", compresslevel=6) as gzip_buffer:
+        # The compression level 6 gives 10% speed gain vs. 2% extra size, in contrast to default compresslevel=9 
         dump_json(data, gzip_buffer)
     base64_bytes = base64.b64encode(buffer.getvalue())
     return base64_bytes.decode("ascii")