- [x] This comment contains a description of changes (with reason)

This is a tradeoff:

Before:
```
           multiqc | Run took 36.22 seconds
           multiqc |  - 6.54s: Searching files
           multiqc |  - 12.93s: Running modules
           multiqc |  - 12.07s: Compressing report data

$ wc -c multiqc_report.html 
34024212 multiqc_report.html
```

After:
```
           multiqc | Run took 32.38 seconds
           multiqc |  - 6.37s: Searching files
           multiqc |  - 13.60s: Running modules
           multiqc |  - 7.64s: Compressing report data

$ wc -c multiqc_report.html 
34522536 multiqc_report.html
```

For comparison, v1.21, using lz compression: 
```
35952091 multiqc_report.html
```

So a report that is less than 2% bigger, for a significant gain in speed. 

The reason I chose 9 initially is because it is the python default and that back then the compression time was not a too big part of the total runtime when the switch to gzip was made. Well things have changed I guess :sweat_smile: .